Rename halftone to halfstep
I was caught between semitone and halfstep here.

I've gone with halfstep because the 'step' helps enforce the
idea of incrementation over the face of the harp.

Emily just thinks I'm trying to be American to be 'different'. Half the
western musical world is American love, it's not 'different', it's
established and consistent.