Complete previous commit
`Position` is indeed overloaded. The candidate fix for this commit was
to successfully overload it in the getter function signature which is
changed in this commit.

I am not entirely comfortable overloading the `Position` name like this.
I won't need the native definition, but it's possible that more issues
like this one could come up. I just need to decide how distracting those
will be vs how distracting a different name would be.