Revert "Convert InteractionMatrix to a 1d mask"
This reverts commit 83156d311bc35dee108afbafe0a019559c88e1cb.

I was wrong. A given row on a harmonica will turn from overblows to blow
bends at some point, and from bends to overdraws at some point. We
either need a matrix to represent that or some kind of encoding of the
boundary where that shift occurs. Matrix is the simpler concept so let's
go with that for now.