Revert "Add sort to the activeIds lists"
This reverts commit 16a749464bfda576468b53fd594c6dd0e31bb370.

The reason being that I determine it to have been a misjudgement of the
correct place to apply the ability to order the ActiveIds lists.

The reducers are complex enough, and keeping their test cases as simple
as possible helps make the intention clear. Returning the deduplicated
row of ActiveIds is sufficient input from the reducers.

The need to order the lists came when I applied more complex matrices to
the inputs, but these complex matrices were not being fed in to the test
cases for this reducer initially, they were being fed in to the
getActiveIds function 2 levels higher (at the time of writing). It seems
much more appropriate (and a tiny bit more efficient) that this function
perform the sorting.

The next commit will introduce the sorting at the getActiveIds function
and the existing test cases at that level will already be set up for it.