Sort the 'other' list in ActiveIds too
The previous commit had us sorting the list which has been generated
from the child reducers. This commit sorts the input, unmodified list
too for consistency. This sort will almost always be redundant, but only
because it's there in the first place. The calling library will usually
be working with lists which are given by this library so now that they
are always ordered going out, they will almost certainly already be
ordered coming back in.