Reduce module depth by replacing child with sibling relationships
This is a trade off between clarity of what is in a module and clarity
of what is exported at each layer. I prefer getting a good impression of
what exists in the moduled and having each file have cleaner and more
similar paths up the hierarchy, and then relying on the root index.ts to
identify what is actually exported. You need to look in the root
index.ts anyway since not everything in a module will be exported
anyway.