Remove unnecessary conditional logic
This relys on the generic function `reverseFromOrigin` to be able to
handle whatever is passed in to it. However I'm having to type cast to
ReadonlyArray<DegreeIds> to prevent it from being invalid. I think the
concept of type narrowing is relevant here. Obviously this is a
workaround and can't stand until the completion of this pull request.