Initial commit of alternate general id ordering module
The OrderedIds module might not be approaching the problem the right
way. It's not possible to indicate in the typing that the type of the
return value is tied to the type of the input value. That makes type
narrowing difficult for the compiler and means it's looking like a
problem to actually use.