Move Covariant typeguards to their own folder
Away from types, because I think it's confusing to have functions in the
types folder.

I was previously exporting the typeguards as types which I think would
have meant that they couldn't be used, and means that they aren't being
used anywhere, but in the end I think it's better to have them exported
and available should the consumer need them.