Reorder the PozitionIds
This requires the creation of another approach to looking up the
pozitions by offset, which is fine, but I'm not entirely happy with the
implementation in getPozitionByOffset. Having there be a possibility of
undefined being returned from the Map and having to deal with that with
a throw so that the undefined type does carry up the call stack makes me
feel that there's something missing in the typing. If we were using
literals for example then we could guarantee that we would be able to
return a value.