Create literal pozition map
This improves the type certainty during key lookup which is a very good
thing. I just need to see whether it can fulfill all of the other
requirements of this map.

I have added a whole load of specialist types here which are probably
not necessary, they just formed part of another idea I was trying out,
and I didn't want to remove it again in case it turned out it was
important. I'll remove it soon enough if I can.