Remove build process and set index.ts as main
The build process might be useful if I want to continue publishing the
package as an npm package in the future. Right now the priority is
getting the package to work in a workspaces context and then a monorepo
so I'm going to use the fact that I know everything about the consuming
packages including the fact that they are happy to transpile the *.ts
files.