Run prettier
And fix a couple of linting errors it created manually.

These were just some double quotes it had felt were necessary because
the strings had \' in them. It's preferred action there is to replace
the \' with ' and replace the surrounding ' with ".

I just removed the \' entirely. Be done with it.