Record changes in CHANGELOG.md
This project can now successfully be run in a yarn workspaces context.
It no longer needs to be published to npm. It might be nice to
reintroduce that feature for the benefit of anyone who might want to
experiment with it, but I really have no evidence that anyone else
cares.

The next step will be to see if the various components of this yarn
workspaces setup and be ported in to a lerna monorepo. If it can then
these projects can all be archived.