Remove npm publishing hooks
We will not be publishing this package for a while. The focus now is to
use it well as part of workspaces and hopefully a monorepo. We may
return to publishing it, but it will probably be a couple of major
versions on by then.

If I had a community of users then I'd feel bad. But I highly doubt I
do. If they complain in the end then I'll give them decent upgrade
instructions for the version that eventually emerges on to npm.