Update package.json version ahead of tag creation
I'm going to be tagging a non-master branch here, because this is a real
landmark position, to be recorded. It is intended to be the point from
which the code in this project is no longer used from here. It will be
part of a monorepo which will be it's own project with all of the
commits imported.

I will archive this project and reference it well from the monorepo. The
branch on which this tag has been created will remain unmerged though,
to preserve the master branch in it's non-workspace context form. If
the monorepo goes well then there will be no need to kick this project
off again, but if it goes badly I may not want to have any of this work
on the master branch at all.